# loader: do not invoke llc separately