--- a/pkg/datapath/loader/loader_test.go
+++ b/pkg/datapath/loader/loader_test.go
@@ -36,10 +36,9 @@ var (
 	contextTimeout = 10 * time.Second
 	benchTimeout   = 5*time.Minute + 5*time.Second
 
-	dirInfo *directoryInfo
-	ep      = testutils.NewTestEndpoint()
-	hostEp  = testutils.NewTestHostEndpoint()
-	bpfDir  = filepath.Join("..", "..", "..", "bpf")
+	ep     = testutils.NewTestEndpoint()
+	hostEp = testutils.NewTestHostEndpoint()
+	bpfDir = filepath.Join("..", "..", "..", "bpf")
 )
 
 // SetTestIncludes allows test files to configure additional include flags.
@@ -56,25 +55,13 @@ func (s *LoaderTestSuite) SetUpSuite(c *C) {
 
 	node.SetTestLocalNodeStore()
 
-	tmpDir, err := os.MkdirTemp("/tmp/", "cilium_")
-	if err != nil {
-		c.Fatalf("Failed to create temporary directory: %s", err)
-	}
-	dirInfo = getDirs(tmpDir)
-
 	cleanup, err := prepareEnv(&ep)
 	if err != nil {
 		SetTestIncludes(nil)
-		os.RemoveAll(tmpDir)
 		c.Fatalf("Failed to prepare environment: %s", err)
 	}
 
-	s.teardown = func() error {
-		if err := cleanup(); err != nil {
-			return err
-		}
-		return os.RemoveAll(tmpDir)
-	}
+	s.teardown = cleanup
 
 	SetTestIncludes([]string{
 		fmt.Sprintf("-I%s", bpfDir),
@@ -135,12 +122,12 @@ func prepareEnv(ep *testutils.TestEndpoint) (func() error, error) {
 	return cleanupFn, nil
 }
 
-func getDirs(tmpDir string) *directoryInfo {
+func getDirs(tb testing.TB) *directoryInfo {
 	return &directoryInfo{
 		Library: bpfDir,
 		Runtime: bpfDir,
 		State:   bpfDir,
-		Output:  tmpDir,
+		Output:  tb.TempDir(),
 	}
 }
 
@@ -150,7 +137,7 @@ func (s *LoaderTestSuite) testCompileAndLoad(c *C, ep *testutils.TestEndpoint) {
 	stats := &metrics.SpanStat{}
 
 	l := NewLoader()
-	err := l.compileAndLoad(ctx, ep, dirInfo, stats)
+	err := l.compileAndLoad(ctx, ep, getDirs(c), stats)
 	c.Assert(err, IsNil)
 }
 
@@ -184,6 +171,7 @@ func (s *LoaderTestSuite) TestReload(c *C) {
 	ctx, cancel := context.WithTimeout(context.Background(), contextTimeout)
 	defer cancel()
 
+	dirInfo := getDirs(c)
 	err := compileDatapath(ctx, dirInfo, false, log)
 	c.Assert(err, IsNil)
 
@@ -221,7 +209,7 @@ func (s *LoaderTestSuite) testCompileFailure(c *C, ep *testutils.TestEndpoint) {
 	var err error
 	stats := &metrics.SpanStat{}
 	for err == nil && time.Now().Before(timeout) {
-		err = l.compileAndLoad(ctx, ep, dirInfo, stats)
+		err = l.compileAndLoad(ctx, ep, getDirs(c), stats)
 	}
 	c.Assert(err, NotNil)
 }
@@ -243,6 +231,8 @@ func BenchmarkCompileOnly(b *testing.B) {
 	ctx, cancel := context.WithTimeout(context.Background(), benchTimeout)
 	defer cancel()
 
+	dirInfo := getDirs(b)
+
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		if err := compileDatapath(ctx, dirInfo, false, log); err != nil {
@@ -258,6 +248,7 @@ func BenchmarkCompileAndLoad(b *testing.B) {
 	defer cancel()
 
 	l := NewLoader()
+	dirInfo := getDirs(b)
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
@@ -272,6 +263,8 @@ func BenchmarkCompileAndLoad(b *testing.B) {
 func BenchmarkReplaceDatapath(b *testing.B) {
 	ctx, cancel := context.WithTimeout(context.Background(), benchTimeout)
 	defer cancel()
+
+	dirInfo := getDirs(b)
 
 	if err := compileDatapath(ctx, dirInfo, false, log); err != nil {
 		b.Fatal(err)