@lmb 2023-11-28
Cleanup is https://github.com/cilium/cilium/pull/29459