@lmb 2023-11-29
I'm not sure actually, clang seems to have had `-mattr` and `-mcpu` since 2005. Maybe something to do with passing `-march`?