@lmb 2023-11-29
Marking as `dont-merge` while I investigate a CI flake.